// schema.d.ts
export interface components {
    schemas: {
// ...
        ScoreDetailAllianceModel_2023: {
            initTeamProp1?: boolean;
            initTeamProp2?: boolean;
            robot1Auto?: boolean;
            robot2Auto?: boolean;
            spikeMarkPixel1?: boolean;
            spikeMarkPixel2?: boolean;
            targetBackdropPixel1?: boolean;
            targetBackdropPixel2?: boolean;
            /** Format: int32 */
            autoBackdrop?: number;
            /** Format: int32 */
            autoBackstage?: number;
            /** Format: int32 */
            dcBackdrop?: number;
            /** Format: int32 */
            dcBackstage?: number;
            /** Format: int32 */
            mosaics?: number;
            /** Format: int32 */
            maxSetLine?: number;
            egRobot1?: components["schemas"]["EndGameLocation"];
            egRobot2?: components["schemas"]["EndGameLocation"];
            /** Format: int32 */
            drone1?: number;
            /** Format: int32 */
            drone2?: number;
            /** Format: int32 */
            minorPenalties?: number;
            /** Format: int32 */
            majorPenalties?: number;
            /** Format: int32 */
            autoNavigatingPoints?: number;
            /** Format: int32 */
            autoRandomizationPoints?: number;
            /** Format: int32 */
            autoBackstagePoints?: number;
            /** Format: int32 */
            autoBackdropPoints?: number;
            /** Format: int32 */
            dcBackdropPoints?: number;
            /** Format: int32 */
            dcBackstagePoints?: number;
            /** Format: int32 */
            mosaicPoints?: number;
            /** Format: int32 */
            setBonusPoints?: number;
            /** Format: int32 */
            egLocationPoints?: number;
            /** Format: int32 */
            egDronePoints?: number;
            /** Format: int32 */
            autoPoints?: number;
            /** Format: int32 */
            dcPoints?: number;
            /** Format: int32 */
            endgamePoints?: number;
            /** Format: int32 */
            penaltyPointsCommitted?: number;
            /** Format: int32 */
            prePenaltyTotal?: number;
            /** Format: int32 */
            totalPoints?: number;
            alliance?: string | null;
            /** Format: int32 */
            team?: number;
        };
        /** IntoTheDeep Alliance Score Breakdown */
        ScoreDetailAllianceModel_2024: {
            robot1Auto?: components["schemas"]["AutoLocation"];
            robot2Auto?: components["schemas"]["AutoLocation"];
            /** Format: int32 */
            autoSampleNet?: number;
            /** Format: int32 */
            autoSampleLow?: number;
            /** Format: int32 */
            autoSampleHigh?: number;
            /** Format: int32 */
            autoSpecimenLow?: number;
            /** Format: int32 */
            autoSpecimenHigh?: number;
            /** Format: int32 */
            teleopSampleNet?: number;
            /** Format: int32 */
            teleopSampleLow?: number;
            /** Format: int32 */
            teleopSampleHigh?: number;
            /** Format: int32 */
            teleopSpecimenLow?: number;
            /** Format: int32 */
            teleopSpecimenHigh?: number;
            robot1Teleop?: components["schemas"]["TeleopLocation"];
            robot2Teleop?: components["schemas"]["TeleopLocation"];
            /** Format: int32 */
            minorFouls?: number;
            /** Format: int32 */
            majorFouls?: number;
            /** Format: int32 */
            autoSamplePoints?: number;
            /** Format: int32 */
            autoSpecimenPoints?: number;
            /** Format: int32 */
            teleopSamplePoints?: number;
            /** Format: int32 */
            teleopSpecimenPoints?: number;
            /** Format: int32 */
            teleopParkPoints?: number;
            /** Format: int32 */
            teleopAscentPoints?: number;
            /** Format: int32 */
            autoPoints?: number;
            /** Format: int32 */
            teleopPoints?: number;
            /** Format: int32 */
            endGamePoints?: number;
            /** Format: int32 */
            foulPointsCommitted?: number;
            /** Format: int32 */
            preFoulTotal?: number;
            /** Format: int32 */
            totalPoints?: number;
            alliance?: string | null;
            /** Format: int32 */
            team?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}